+++ Editor/AssemblyInfo.cs
@@ -1,4 +1,3 @@
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("GitPackageManager.Editor.Tests")]
-[assembly: InternalsVisibleTo("Essentials.GitPackageManager.Editor.Tests")]
+[assembly: InternalsVisibleTo("MartinCalander.GitPackageManager.Editor.Tests")]