+++ Editor/PackageManagerWindow.Details.cs
@@ -3,7 +3,7 @@ using System.IO;
 using UnityEditor;
 using UnityEngine;
 
-namespace Essentials.GitPackageManager.Editor
+namespace MartinCalander.GitPackageManager.Editor
 {
     public partial class GitPackageManagerWindow
     {