+++ Editor/PackageManagerWindow.Popups.cs
@@ -1,7 +1,7 @@
 using UnityEditor;
 using UnityEngine;
 
-namespace Essentials.GitPackageManager.Editor
+namespace MartinCalander.GitPackageManager.Editor
 {
     public partial class GitPackageManagerWindow
     {