+++ Editor/Utilities/CliInstaller.cs
@@ -1,7 +1,7 @@
 using System;
 using UnityEngine;
 
-namespace Essentials.GitPackageManager.Editor
+namespace MartinCalander.GitPackageManager.Editor
 {
     internal sealed class CliInstallPlan
     {